Leave it at here for the moment
For the time being, leave this as the last commit for CVNP, so changes can be pulled and fresh developmet be done elsewhere. It is possible that testing for CVNP may be easier with hardware and an oscilloscope than through a software simulation, so that will be attempted, and this can optionally be revisited later.